search feature

+++ src/main/java/com/backend/elearning/domain/order/OrderVM.java
@@ -11,9 +11,9 @@ public record OrderVM(
         String createdAt,
         String status,
         List<OrderDetailVM> orderDetails,
-        Double totalPrice
+        Long totalPrice
 ) {
-    public static OrderVM fromModel(Order order, List<OrderDetailVM> orderDetails, Double total) {
+    public static OrderVM fromModel(Order order, List<OrderDetailVM> orderDetails, Long total) {
         String student = order.getStudent().getEmail();
         String coupon = order.getCoupon() != null ? order.getCoupon().getCode() : "No coupon";
         return new OrderVM(order.getId(),student, coupon, order.getCreatedAt().toString(), order.getStatus().toString(), orderDetails, total);
